demo/camera-server/requirements.txt
paho-mqtt
Pillow
